pyproject.toml
[project]
name = "senaiapi"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.9"
dependencies = [
    "bcrypt>=5.0.0",
    "fastapi>=0.128.8",
    "httpx>=0.28.1",
    "huggingface-hub>=1.8.0",
    "limits>=4.2",
    "passlib[bcrypt]>=1.7.4",
    "pdfplumber>=0.11.8",
    "polars>=1.36.1",
    "psycopg2-binary>=2.9.12",
    "pymupdf>=1.26.5",
    "pytest>=8.4.2",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "slowapi>=0.1.9",
    "sqlalchemy>=2.0.49",
    "uvicorn[standard]>=0.39.0",
]
